requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
openai==1.3.5
supabase==2.0.3
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.0.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6 